[Fix] CI.
